requirements.txt
python-telegram-bot==21.9 
discord.py==2.4.0
python-dotenv==1.0.1 
request==2.32.3 